requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SocketIO==5.3.6
WTForms==3.0.1
email-validator==2.0.0
qrcode==7.4.2
Pillow>=9.5.0,<10.0.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
